+++ common/src/main/java/jimenezli/neuro21/entity/GigaVedalEntity.java
@@ -1,9 +1,9 @@
 package jimenezli.neuro21.entity;
 
 import net.minecraft.world.entity.EntityType;
-import net.minecraft.world.entity.Mob;
 import net.minecraft.world.entity.ai.attributes.AttributeSupplier;
 import net.minecraft.world.entity.ai.attributes.Attributes;
+import net.minecraft.world.entity.animal.Turtle;
 import net.minecraft.world.level.Level;
 
 public class GigaVedalEntity extends VedalEntity {
@@ -12,11 +12,11 @@ public class GigaVedalEntity extends VedalEntity {
     }
 
     public static AttributeSupplier.Builder createAttributes() {
-        return Mob.createMobAttributes().add(Attributes.ATTACK_DAMAGE, 2.0D).add(Attributes.MOVEMENT_SPEED, 0.3).add(Attributes.MAX_HEALTH, 20.0);
+        return Turtle.createAttributes().add(Attributes.ATTACK_DAMAGE, 2.0D).add(Attributes.MOVEMENT_SPEED, 0.3).add(Attributes.MAX_HEALTH, 40.0);
     }
 
     public void tick() {
         super.tick();
-        this.setSpeed(0.3f);
+        this.setSpeed(1.0f);
     }
 }