+++ common/src/main/java/jimenezli/neuro21/util/EntityNames.java
@@ -8,7 +8,6 @@ public class EntityNames {
     public static final String NEUROSAMA = "neurosama";
     public static final String EVIL_NEUROSAMA = "evil_neurosama";
     public static final String HIYORI = "hiyori";
-    public static final String IRON_COW = "iron_cow";
     public static final String HIYORI_BOSS = "hiyori_boss";
     public static final String GYMBAG = "gymbag";
     public static final String SWARM_DRONE = "swarm_drone";